+++ aksel.nav.no/website/components/sanity-modules/props/PropTabell.tsx
@@ -1,15 +1,7 @@
 import { withErrorBoundary } from "@/error-boundary";
-import { SanityT } from "@/lib";
 import { BodyShort, Detail, Heading, Tooltip } from "@navikt/ds-react";
 import { Highlighter } from "./Highlight";
-
-export type PropT = {
-  _type: "komponent";
-  _key: string;
-  title?: string;
-  overridable?: boolean;
-  propref?: SanityT.Schema.ds_props;
-};
+import { PropTableT } from "@/types";
 
 const List = ({ prop, parent }: { prop: any; parent: string }) => {
   if (prop?.description && prop.description.includes("@private")) {
@@ -55,7 +47,7 @@ const List = ({ prop, parent }: { prop: any; parent: string }) => {
   );
 };
 
-const PropTable = ({ komponent }: { komponent: PropT }) => {
+const PropTable = ({ komponent }: { komponent: PropTableT }) => {
   return (
     <div>
       <Heading