+++ @navikt/core/react/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @navikt/ds-react
 
+## 2.8.14
+
+### Patch Changes
+
+- [`f0d9a8853`](https://github.com/navikt/aksel/commit/f0d9a8853f56854a4049bd3f6cc34968e9d6c380) Thanks [@KenAJoh](https://github.com/KenAJoh)! - Fixes shown month on open in Datepicker/MonthPicker, better handling of out-of-range 'today'
+
+- Updated dependencies []:
+  - @navikt/aksel-icons@2.8.14
+  - @navikt/ds-icons@2.8.14
+
+## 2.8.13
+
+### Patch Changes
+
+- Updated dependencies [[`ecd6241e5`](https://github.com/navikt/aksel/commit/ecd6241e5751519fbda800a5f41a613957dbf77c)]:
+  - @navikt/aksel-icons@2.8.13
+  - @navikt/ds-icons@2.8.13
+
 ## 2.8.12
 
 ### Patch Changes